migrate to use runtime_data
