+++ homeassistant/components/zwave_me/climate.py
@@ -11,13 +11,13 @@ from homeassistant.components.climate import (
     ClimateEntityFeature,
     HVACMode,
 )
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import ATTR_TEMPERATURE
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from .const import DOMAIN, ZWaveMePlatform
+from . import ZWaveMeConfigEntry
+from .const import ZWaveMePlatform
 from .entity import ZWaveMeEntity
 
 TEMPERATURE_DEFAULT_STEP = 0.5
@@ -27,7 +27,7 @@ DEVICE_NAME = ZWaveMePlatform.CLIMATE
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    config_entry: ConfigEntry,
+    config_entry: ZWaveMeConfigEntry,
     async_add_entities: AddConfigEntryEntitiesCallback,
 ) -> None:
     """Set up the climate platform."""
@@ -35,7 +35,7 @@ async def async_setup_entry(
     @callback
     def add_new_device(new_device: ZWaveMeData) -> None:
         """Add a new device."""
-        controller = hass.data[DOMAIN][config_entry.entry_id]
+        controller = config_entry.runtime_data
         climate = ZWaveMeClimate(controller, new_device)
 
         async_add_entities(
